refactor: move teamrbac metrics into the internal metrics package
On-behalf-of: @SAP krzysztof.zagorski@sap.com
