+++ internal/controller/teamrbac/teamrolebinding_controller.go
@@ -31,6 +31,7 @@ import (
 	greenhousev1alpha2 "github.com/cloudoperators/greenhouse/api/v1alpha2"
 	"github.com/cloudoperators/greenhouse/internal/clientutil"
 	"github.com/cloudoperators/greenhouse/internal/lifecycle"
+	"github.com/cloudoperators/greenhouse/internal/metrics"
 	"github.com/cloudoperators/greenhouse/internal/util"
 )
 
@@ -97,7 +98,7 @@ func (r *TeamRoleBindingReconciler) setConditions() lifecycle.Conditioner {
 		readyCondition := computeReadyCondition(trb.Status)
 		ownerLabelCondition := util.ComputeOwnerLabelCondition(ctx, r.Client, trb)
 		trb.Status.SetConditions(readyCondition, ownerLabelCondition)
-		updateMetrics(trb)
+		metrics.UpdateTeamrbacMetrics(trb)
 	}
 }
